The query string is passed directly to generators.
prefs.py no longer has a dict of dicts - it has a dict of functions.
Take a look at the example prefs.py for the new format.  If you relied
on something special being passed to your generator write a decorator
in prefs.py or write your generator to get what it needs out of the
query string.

This requires you to fix some feed URLs but lets you add new feeds
quickly and without editing prefs.py.

The twitter_noreplies generator now looks up twitter ids by itself.  It
takes the twitter username string as a single arg.  This means that the
query string must be cased exactly as the username is on twitter.com.