Pass dicts to generators instead of strings.
Prefs are bumped from version 3 to version 4.  Unfortunately if you were
manipulating the argument string in a fuction in prefs.py you'll have to
rewrite chunks of it.

There are now old-style query strings and new-style query strings.
Old-style queries have the feed=foo key like you're used to using.
Old-style queries will be parsed into dicts behind the scenes but won't
be able to take advantage of extra features.

New-style queries have two mandatory args in the query string: gen and
arg.  gen is the same as the generator name defined in prefs.py.  arg is
the "primary argument" and is the same as what you'd put after the
underscore in the old-style query.

Generators are now passed a dict.  The changes to the generator spec
will be updated in a later documentation patch.  The argument dict has a
"qs" sub-dict which contains the query string's key/value pairs.