Make err() more noisy
create_error_feed()'s IDs used to be constant across all feeds and
errors.  A timestamp is now appended to the ID so a conforming RSS
reader will show a new entry each time a fetch results in an error.

There's also an optional silent arg to err() which suppresses output.